+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/web/ProjectController.java
@@ -142,6 +142,7 @@ public class ProjectController {
             .model(InvoicingProject.class.getName())
             .add("form", "invoicing-project-form")
             .param("forceEdit", "true")
+            .param("show-toolbar", "false")
             .context("_project", project)
             .map());
   }